Add the armv5e architecture to binutils
Binutils has a comment that indicates it is supposed to match gcc for
all of the support "-march=" settings, but it was lacking the armv5e setting.
This was a simple way to add it, as thumb instructions shouldn't be generated
by the compiler anyway.

Upstream-Status: Denied
Upstream maintainer indicated that we should not be using armv5e, even
though it is a legal archicture defined by our gcc.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>